add serve cog endpoint
